+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Fixed `Data.add_data_frame()` in order to handle `None` and `pd.Series` as `data_frame` parameter.
 
+### Added
+
+- Separated the JavaScript code into a JavaScript file instead of the previous string template.
+
 ## [0.7.0] - 2022-03-29
 
 ### Fixed